Sketcher: Use the same workflow as Draw for Drag
